+++ src/interface/src/app/treatments/printable-treatment/printable-treatment.component.scss
