+++ src/interface/src/app/treatments/treatment-config/treatment-config.component.html
@@ -53,15 +53,8 @@
       *ngIf="(showTreatmentLegend$ | async) || false"></app-treatment-legend>
   </div>
 </section>
-<div
-  id="printable-map"
-  style="
-    position: absolute;
-    width: 1000px;
-    height: 600px;
-    left: -9999px;
-    top: -9999px;
-  "></div>
-<app-printable-tx-footer
+
+<app-printable-treatment
   id="pdffooter"
-  [summary]="summary$ | async"></app-printable-tx-footer>
+  (createPdf)="createPDF()"
+  [summary]="summary$ | async"></app-printable-treatment>
